refactor: トップページを Server/Client 分割し爆発を framer-motion 化
frontend.md のルール（page.tsx は Server Component / 'use client' は末端限定）
に沿って分割。見た目・挙動は変えない純粋リファクタ。

- page.tsx: 'use client' を外し Server Component 化、組み立てのみに
- TopMenu.tsx(Client): ボタン・モーダル・爆発・SE・遷移を集約
- MbtiDecorations.tsx(Server): MBTIキャラ4枚を配列化(DRY)
- SparklesExplosion.tsx: 爆発を inline style/@keyframes から framer-motion へ置換

refs #192
